SSR-18-1 Implement admin auth page and logic: add admin auth endpoints
