+++ modules/reactmono/admin-auth/index.js
@@ -1,5 +1,4 @@
 require = require('esm')(module, {await: true});
-module.exports.getRoutes = require('./src/getRoutes').default;
-module.exports.models = {
-    admins: require('./src/model/Admin').default
-};
+module.exports.getRoutes = require('./src/route').default;
+module.exports.models = require('./src/model').default;
+module.exports.middleware = require('./src/middleware').default;
