+++ modules/reactmono/admin-auth/src/middleware/auth.js
@@ -1,0 +1,22 @@
+import jwt from 'jsonwebtoken';
+import config from 'config';
+
+export default (req, res, next) => {
+    // Get token from header
+    const token = req.header('x-auth-token');
+
+    // Check if not token
+    if (!token) {
+        return res.status(401).json({msg: 'No token, authorization denied'});
+    }
+
+    // Verify token
+    try {
+        const decoded = jwt.verify(token, config.get('jwtSecret'));
+
+        req.admin = decoded.admin;
+        next();
+    } catch (err) {
+        res.status(401).json({ msg: 'Token is not valid' });
+    }
+};
