+++ modules/reactmono/admin-auth/src/middleware/index.js
@@ -1,0 +1,5 @@
+import auth from './auth';
+
+export default {
+    auth
+}
