+++ modules/reactmono/admin-auth/src/model/index.js
@@ -1,0 +1,5 @@
+import Admin from './Admin';
+
+export default {
+    admins: Admin
+};
