+++ modules/reactmono/admin-auth/src/route/index.js
@@ -1,0 +1,12 @@
+import signin from './signin';
+import signup from './signup';
+import me from './me';
+
+/**
+ * Admin Authentication Routes.
+ */
+export default () => ([
+    signin(),
+    signup(),
+    me()
+]);
